+++ src/common/Event.test.ts
@@ -6,15 +6,28 @@ import { Event, EventEmitter, MemorizingEventEmitter } from './Event';
 
 describe('Event', () => {
   it('emits events', () => {
-    const s = stub();
+    const s1 = stub();
+    const s2 = stub();
+    const s3 = stub();
     const emitter = new EventEmitter<number>();
 
-    const l = emitter.addListener(s);
-    emitter.emit(42);
-    l.dispose();
-    emitter.emit(43);
+    const l1 = emitter.addListener(s1);
+    emitter.emit(1);
+    const l2 = emitter.addListener(s2);
+    emitter.emit(2);
+    const l3 = emitter.addListener(s3);
+    emitter.emit(3);
 
-    expect(s).to.have.been.calledOnceWith(42);
+    l1.dispose();
+    emitter.emit(4);
+    l2.dispose();
+    emitter.emit(5);
+    l3.dispose();
+    emitter.emit(6);
+
+    expect(s1.args).to.deep.equal([[1], [2], [3]]);
+    expect(s2.args).to.deep.equal([[2], [3], [4]]);
+    expect(s3.args).to.deep.equal([[3], [4], [5]]);
   });
 
   it('memorizes event emissions', () => {
@@ -60,7 +73,7 @@ describe('Event', () => {
     emitter.emit(42);
     expect(await v).to.equal(42);
 
-    expect((emitter as any).listeners.size).to.equal(0);
+    expect(emitter.size).to.equal(0);
   });
 
   it('cancels conversion to promise', async () => {
@@ -69,13 +82,13 @@ describe('Event', () => {
     setTimeout(() => cts.cancel(), 1);
     const v = Event.toPromise(emitter.addListener, cts.token);
     await expect(v).to.eventually.be.rejectedWith(TaskCancelledError);
-    expect((emitter as any).listeners.size).to.equal(0);
+    expect(emitter.size).to.equal(0);
   });
 
   it('cancels conversion to promise sync', async () => {
     const emitter = new EventEmitter<number>();
     const v = Event.toPromise(emitter.addListener, CancellationToken.Cancelled);
     await expect(v).to.eventually.be.rejectedWith(TaskCancelledError);
-    expect((emitter as any).listeners.size).to.equal(0);
+    expect(emitter.size).to.equal(0);
   });
 });